fix: provide solution that satisfies requirement
The specification says we have to return something that implements
Animal, but can't tell which at compile time.  The existing solution
only returns one concrete type Sheep, while the intent was to return
either Sheep or Cow.  This requires type erasure that can only be done
with a reference to dyn trait.